feat: add island entry and dark mode module
