+++ assets/js/dark-mode.js
@@ -1,0 +1,16 @@
+export function initDarkMode() {
+  const darkModeToggle = document.getElementById("dark-mode-toggle");
+  if (!darkModeToggle) return;
+
+  if (localStorage.getItem("darkMode") === "true") {
+    document.body.classList.add("dark-mode");
+  }
+
+  darkModeToggle.addEventListener("click", () => {
+    document.body.classList.toggle("dark-mode");
+    localStorage.setItem(
+      "darkMode",
+      document.body.classList.contains("dark-mode")
+    );
+  });
+}
