+++ index.html
@@ -32,8 +32,8 @@
   <footer aria-label="Project contribution">
     <p><a href="CONTRIBUTING.md">Contribute to this project</a></p>
   </footer>
-  <script src="script.js"></script>
-  <script src="assets/js/app.js"></script>
+    <script src="assets/js/islands.js"></script>
+    <script src="assets/js/app.js"></script>
   <script src="https://unpkg.com/web-vitals@3/dist/web-vitals.iife.js"></script>
   <script src="assets/js/metrics.js"></script>
 </body>
