+++ package.json
@@ -2,7 +2,7 @@
   "name": "cybersecuirtydictionary",
   "version": "1.0.0",
   "description": "https://alex-unnippillil.github.io/CyberSecuirtyDictionary/",
-  "main": "script.js",
+    "main": "assets/js/islands.js",
   "scripts": {
     "build": "node scripts/build.js",
     "test": "html-validate index.html search.html diagnostics.html && node diagnostics.test.js",
